jobs-scraper/requirements.txt
anyio==4.2.0
asyncio==3.4.3
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
black==23.10.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
constantly==15.1.0
cryptography==41.0.4
cssselect==1.2.0
DateTime==5.2
decorator==5.1.1
docker==6.1.3
exceptiongroup==1.2.0
fake-useragent==1.4.0
filelock==3.12.4
h11==0.14.0
html2text==2020.1.16
httpcore==1.0.2
httpx==0.26.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
install==1.3.5
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
joblib==1.3.2
logging==0.4.9.6
lxml==4.9.3
mypy-extensions==1.0.0
mysql-connector-python==8.2.0
packaging==23.2
parsel==1.8.1
pathspec==0.11.2
platformdirs==3.11.0
Protego==0.3.0
protobuf==4.21.12
proxyscrape==0.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==23.2.0
pytz==2023.3.post1
PyYAML==6.0.1
queuelib==1.6.2
regex==2023.12.25
requests==2.31.0
requests-file==1.5.1
retry2==0.9.5
retrying==1.3.4
scrapinghub==2.4.0
Scrapy==2.11.0
scrapy-proxy-pool==0.1.9
scrapy-rotating-proxies==0.6.2
scrapy-user-agents==0.1.1
scrapy-zyte-smartproxy==2.2.0
service-identity==23.1.0
shub==2.14.5
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
tldextract==5.0.1
toml==0.10.2
tomli==2.0.1
tqdm==4.55.1
Twisted==22.10.0
typing==3.7.4.3
typing_extensions==4.8.0
ua-parser==0.18.0
urllib3==2.0.7
user-agents==2.2.0
w3lib==2.1.2
websocket-client==1.6.4
zope.interface==6.1
